requirements.txt
flask>=3.0.0
google-genai>=1.0.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
flask-limiter>=3.5.0
google-cloud-logging>=3.8.0
pytest>=8.0.0
